requirements.txt
Pillow
huggingface_hub
openai
torch
plotly
requests